Rename Hammer Time to Hammertime.
